ToWkt for non-geotypes
This is primarily just moving code around - no changes to
implementation, though I did rename the `conversion` module to
`geo_types_from_wkt` and deprecated the old name.
